Added:    @option enable-printf / @option enable-println and    @printf(...); / @println(...);
Both options for each addition are interchangable regardless of the language of the parser.

While developing a parser it may be advantageous to sprinkle printf messages into the grammar, so that flow of looking for various token types can be followed without having to single step through the parsing. Previously this could be done easily enough with the double brace syntax, {{ printf(...); }}, but this requires either that debug printfs remain live in the grammar or be physically removed. The @printf() meta allows for a printf to remain in the grammar but lay dormant; it is ignored if @option enable-printf; is not present.