Added a new method to the parser: - (BOOL)parseString:(NSString *)string asRule:(NSString *)ruleName result:(id *)result;
This method is intended to be used for debugging purposes. In normal use, failures in parsing yield error messages that are difficult to get value from if the input is large. This is because all you really know is the top level rule did
not parse.

Using this new method, you can easily iterate through various rules based on the
input you are providing. This allows you to more easily identify which rule is not working according to your expectations.