Added @interface, @implementation, and @end directives, and modified the @init directive.
Previously the syntax for @init was:
   @init {
   ...
   }
and now its:
   @init
   ...
   @end

Syntax for @implementation and @interface are the same, but these permit optional text to appear on the same line following the opening directive. Any text on that line is ignored, and the rest of the content until the @end directive is placed into the generated parser's implementation or interface accordingly.